Starting dao implementation of the interface's model.
Need to develop a persistence mode in order to use the CRUD
algorithm by files with a database. By the way, have to finish
the CRUD with files mode that I have already started.

The create() and read() functions are already finish by the file
mode. Design how to do the database.